Fixes to initializer and converter and some other stuff

+++ src/orbconverter.cpp
@@ -109,8 +109,7 @@ cv::Mat Orbconverter::toCvSE3(const Eigen::Matrix3d &R, const Eigen::Vector3d &t
 Eigen::Vector3d Orbconverter::toVector3d(const cv::Mat &cvVector)
 {
     Eigen::Vector3d vector;
-    vector << cvVector.at<double>(0), cvVector.at<double>(1), cvVector.at<double>(2);
-
+    vector << cvVector.at<float>(0), cvVector.at<float>(1), cvVector.at<float>(2);
     return vector;
 }
 
@@ -126,9 +125,9 @@ Eigen::Matrix3d Orbconverter::toMatrix3d(const cv::Mat &cvMat3)
 {
     Eigen::Matrix3d matrix;
 
-    matrix << cvMat3.at<double>(0,0), cvMat3.at<double>(0,1), cvMat3.at<double>(0,2),
-         cvMat3.at<double>(1,0), cvMat3.at<double>(1,1), cvMat3.at<double>(1,2),
-         cvMat3.at<double>(2,0), cvMat3.at<double>(2,1), cvMat3.at<double>(2,2);
+    matrix << cvMat3.at<float>(0,0), cvMat3.at<float>(0,1), cvMat3.at<float>(0,2),
+         cvMat3.at<float>(1,0), cvMat3.at<float>(1,1), cvMat3.at<float>(1,2),
+         cvMat3.at<float>(2,0), cvMat3.at<float>(2,1), cvMat3.at<float>(2,2);
 
     return matrix;
 }
